Medicine field + schedule update
Checks to see if fields are empty and if so makes them null. It wont send to database if no name in medicine textbox and alerts user such.

Made it so empty fields in database also doesn't show up as undefined in schedule and instead just deletes that section on the schedule.